requirements.txt
streamlit
transformers
torch
pillow
wikipedia
requests
spacy
geopy
numpy==1.26.4
